Split OpenSSL backend into separate files (RSA)
